+++ src/lib.rs
@@ -1,5 +1,9 @@
-pub mod message;
-
+mod beat_number;
+mod line_launcher;
+mod message;
 mod result;
 
+pub use beat_number::BeatNumber;
+pub use line_launcher::LineLauncher;
+pub use message::Message;
 pub use result::Result;
